+++ src/panel/state.ts
@@ -1,11 +1,18 @@
-import { State } from './'
+import type { State } from './'
 
 export let state: State
 
-export const initializeState = (initialState: State) => (state = initialState)
+export const initializeState = (initialState: State): void => {
+  state = initialState
+}
 
-export const setState = <K extends keyof State>(key: K, value: State[K]) =>
-  (state = {
+export const setState = <K extends keyof State>(
+  key: K,
+  value: State[K]
+): State => {
+  state = {
     ...state,
     [key]: value,
-  })
+  }
+  return state
+}