+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [1.6.0] - 2023-05-21
+### Added
+- Optional XP counter to display current progress (disabled by default, can be enabled in settings)
+- Improved extension compatibility with Vim and other extensions that use the keyboard
+
 ## [1.1.0] - 2022-07-16
 ### Added
 - New Codachi pet 🐶